pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "case-extraction-tools"
version = "0.2.0"
description = "Convert research docs into AAB Case Registry entries"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "pypdf>=3.0",
    "python-docx>=1.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "reportlab>=4.0",
    "openai>=1.0",
    "python-dotenv>=1.0",
    "jsonschema>=4.0",
    "jinja2>=3.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v"
filterwarnings = ["ignore::DeprecationWarning"]
